test: change test runner to vitest (#525)
* test: change test runner to vitest

This change moves the test runner from `mocha` + `nyc` to `vitest`.

In order to import `vitest` the tests had to be in esm, so I checked out everything under the `tests` folder from https://github.com/eslint-community/eslint-plugin-eslint-plugin/pull/516 (leaving the source as it was).

I also moved the `eslint-rule-tester` out of the `tests/lib` folder and into its own `utils` folder.  That way vitest didn't treat it as a test.

Note: rather than using `vitest`'s preferred `v8` coverage reporter, I used the `istanbul` one, to ensure the coverages between old and new were the same.  I did notice that when I tried the `v8` coverage reporter, the coverage numbers were much less.  Something to consider as a follow-up change.  The v8 reporter *should* be more accurate, so the coverage may not be as high as it seems.  I'd recommend moving up to the `v8` reporter after the esm branch lands.

* test: update fixtures tsconfig to use wildcard

* build: remove `globals` and packageManager config