build: migrate eslint.config.js to typescript
This change migrates the eslint config for the repo to typescript.  This is in support of the larger migration to typescript.  In order for us to continue to consume the plugin locally, after it's been converted to typescript (without having to build first), we'll need to be loading the config from typescript.

Since I needed to update eslint to support, I took the opportunity to all of the other plugins too (except unicorn).

I also removed a rule disable that had a note on it about dropping after we dropped support for node 14.