refactor: remove context compat functions (#532)
This change is in support of the larger migration to TypeScript., and a followup to the removal of support for ESLint v8.  The context compatibility functions in the `utils` module were no longer necessary.  I originally did this in the TypeScript branch, but that PR's going to be large enough.  So, I'm trying to peel off smaller independent changes I can land separately first.