ci: add e2e tests to CI
This change adds a new end-to-end test capability.  There are two new jobs in the main CI workflow.  One job to build the package (which also good to have outside of the need to run e2e tests), executed with the latest LTS node.  The build artifact is uploaded for the downstream `e2e` job to use.  The `e2e` job, downloads the artifact, run `npm install` and then `npm run e2e`, which executes a custom node script to loop through subdirectories of `/e2e`, which each contain a project fixture with different setups.  Both projects have the same simple plugin source code, and each has this plugin (referenced from the root of the repo (aka the result of the build)) and `eslint` installed as `devDependencies`.  Each also has a `lint` package.json script that runs `eslint` on the simple plugin.

\## `/e2e/all`

This fixture has a regular js config and runs our `all` config on the project.

\## `/e2e/all-typed-config`

This fixture has `typescript` installed and uses a ts-based config and runs our `all` config on the project.  It also executes `tsc` as part of the `lint` command, to ensure everything's typed property.