feat(test-case-name-property): add rule to require `name` for test cases
This change adds a new rule to require `name` be present in test cases under certain conditions.  This rule aims to ensure test suites are producing logs in a form that make it easy to identify failing test, when they happen.
For thoroughly tested rules, it's not uncommon to have the same `code` across multiple test cases, with only `options` or `settings` differing between them.
Requiring these test cases to have a `name` helps ensure the test output is meaningful and distinct.